--- src/main/java/com/example/codingmall/Login/OAuth2dto/OAuth2Response.java
@@ -1,18 +1,0 @@
-package com.example.codingmall.Login.OAuth2dto;
-
-public interface OAuth2Response {
-    // 제공자
-    String getProvider();
-
-    // 제공자에서 발급해준 ID
-    String getProviderId();
-
-    // 이메일
-    String getEmail();
-
-    // 사용자 실명
-    String getName();
-
-    // 프로필 이미지
-    String getProfileImage();
-}
