+++ src/main/java/com/example/codingmall/User/Login/Jwt/JWTAuthFilter.java
@@ -1,4 +1,4 @@
-package com.example.codingmall.Login.Jwt;
+package com.example.codingmall.User.Login.Jwt;
 
 import com.example.codingmall.User.UserService;
 import jakarta.servlet.FilterChain;
@@ -7,7 +7,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
